apu: move cpal stream code
